Rename Transaction to TransactionalCallback, move it to sqlobject, and fix its generic signature to refer to Transactional. Add TransactionalConsumer and Transactional.useTransaction(TransactionalConsumer)
