+++ sqlobject/src/main/java/org/jdbi/v3/sqlobject/TransactionalCallback.java
@@ -13,20 +13,18 @@
  */
 package org.jdbi.v3.sqlobject;
 
-import java.util.function.Supplier;
-
-import net.sf.cglib.proxy.MethodProxy;
-
-import org.jdbi.v3.Handle;
+import org.jdbi.v3.TransactionStatus;
 import org.jdbi.v3.sqlobject.mixins.Transactional;
 
-class InTransactionHandler implements Handler
+public interface TransactionalCallback<R, T extends Transactional<T>, X extends Exception>
 {
-    @Override
-    public Object invoke(Supplier<Handle> handle, final Object target, Object[] args, MethodProxy mp) throws Exception
-    {
-        @SuppressWarnings("unchecked")
-        final TransactionalCallback callback = (TransactionalCallback) args[0];
-        return handle.get().inTransaction((conn, status) -> callback.inTransaction((Transactional) target, status));
-    }
+    /**
+     * Execute in a transaction. Will be committed afterwards, or rolled back if an exception is thrown.
+     *
+     * @param transactional The object communicating with the database.
+     * @param status a handle on the transaction, kind of
+     * @return the transaction result
+     * @throws X any exception thrown will cause the transaction to be rolled back
+     */
+    R inTransaction(T transactional, TransactionStatus status) throws X;
 }
